Various  Changes

+++ Armpra/Assets/Scenes/Gameplay/Ingame/Enemies/Scripts/EnemySpawner.cs
@@ -24,7 +24,7 @@ public class EnemySpawner : MonoBehaviour {
         }
         //Outputting Texts
         rf.enemiesRemainingText.text = (RuntimeSpecs.maxEnemyCount - RuntimeSpecs.enemiesKilled).ToString();
-        RuntimeSpecs.ap = (float)Mathf.Round(((float)RuntimeSpecs.enemiesKilled / RuntimeSpecs.maxEnemyCount) * 100f);
+        RuntimeSpecs.ap = Mathf.Round(((float)RuntimeSpecs.enemiesKilled / RuntimeSpecs.maxEnemyCount) * 100f);
         rf.bapText.text = RuntimeSpecs.ap.ToString();
         //Debug.Log("Max:" + RuntimeSpecs.maxEnemyCount + ", Spawned: " + enemiesSpawned + ", Killed: " + RuntimeSpecs.enemiesKilled + ", Bap: " + RuntimeSpecs.bap);
     }
